+++ commitlint.config.js
@@ -3,6 +3,6 @@ module.exports = {
   'type-enum': [
     2,
     'always',
-    ['feat', 'fix', 'docs', 'style', 'refactor', 'test', 'chore'],
+    ['feat', 'fix', 'docs', 'refactor', 'test', 'chore'],
   ],
 };