+++ src/@graphql/graphql.module.ts
@@ -1,10 +1,8 @@
 import { GraphQLModule } from '@nestjs/graphql';
-import ConfigModule from '../@config/config.module';
-import appConfig from '../@config/app.config';
+import ConfigModule from '~/@config/config.module';
 import graphqlFactory from './graphql.factory';
 
 export default GraphQLModule.forRootAsync({
   imports: [ConfigModule],
   useFactory: graphqlFactory,
-  inject: [appConfig.KEY],
 });