+++ src/app.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { APP_GUARD } from '@nestjs/core';
 import { ThrottlerGuard, ThrottlerModule } from '@nestjs/throttler';
+import { UserModule } from './user/user.module';
 import ConfigModule from './@config/config.module';
 import DBModule from './@database/db.module';
 
@@ -12,6 +13,7 @@ import DBModule from './@database/db.module';
       ttl: Number(process.env.THROTTLER_TTL),
       limit: Number(process.env.THROTTLER_LIMIT),
     }),
+    UserModule,
   ],
   providers: [
     {