+++ src/lib/isEnv.ts
@@ -1,4 +1,2 @@
-import { ProcessEnvOptions } from 'node:child_process';
-
 export default (env: 'test' | 'local' | 'dev' | 'staging' | 'prod') =>
   process.env.NODE_ENV === env;